student crud starting

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -30,12 +30,12 @@ class CreateUsersTable extends Migration
       $table->boolean('state')->default(0)->nullable();
       $table->boolean('status')->default(0)->nullable();
       $table->boolean('root')->default(0);
-      $table->integer('school_id')->nullable();
-      $table->integer('roles_id')->nullable();
+      $table->bigInteger('school_id')->nullable();
+      $table->bigInteger('roles_id')->nullable();
       $table->timestamp('email_verified_at')->nullable();
-      $table->integer('created_user')->nullable();
-      $table->integer('updated_user')->nullable();
-      $table->integer('deleted_user')->nullable();
+      $table->bigInteger('created_user')->nullable();
+      $table->bigInteger('updated_user')->nullable();
+      $table->bigInteger('deleted_user')->nullable();
       $table->rememberToken();
       $table->softDeletes();
       $table->timestamps();
