+++ database/migrations/2020_11_18_170119_create_courses_table.php
@@ -15,9 +15,9 @@ class CreateCoursesTable extends Migration
   {
     Schema::create('courses', function (Blueprint $table) {
 
-      $table->integer('teacher_id');
-      $table->integer('student_id');
-      $table->integer('discipline_id');
+      $table->bigInteger('teacher_id');
+      $table->bigInteger('student_id');
+      $table->bigInteger('discipline_id');
       $table->integer('status_teacher');
       $table->integer('status_student');
       $table->string('com_status_teacher', 256)->nullable();
@@ -26,9 +26,9 @@ class CreateCoursesTable extends Migration
       $table->string('customtitle', 100)->nullable();
       $table->time('custom_start')->nullable();
       $table->time('custom_end')->nullable();
-      $table->integer('created_user');
-      $table->integer('updated_user');
-      $table->integer('deleted_user')->nullable();
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
       $table->softDeletes();
       $table->timestamps();
     });
