+++ database/migrations/2020_11_18_171511_create_textbooks_table.php
@@ -6,31 +6,31 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateTextbooksTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('textbooks', function (Blueprint $table) {
-            $table->id();
-            $table->integer('classroom_id');
-            $table->integer('created_user');
-            $table->integer('updated_user');
-            $table->integer('deleted_user')->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
-    }
+  /**
+   * Run the migrations.
+   *
+   * @return void
+   */
+  public function up()
+  {
+    Schema::create('textbooks', function (Blueprint $table) {
+      $table->id();
+      $table->bigInteger('classroom_id');
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
+      $table->softDeletes();
+      $table->timestamps();
+    });
+  }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('textbooks');
-    }
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  public function down()
+  {
+    Schema::dropIfExists('textbooks');
+  }
 }
