+++ database/migrations/2020_12_22_174552_create_paids_table.php
@@ -15,13 +15,13 @@ class CreatePaidsTable extends Migration
   {
     Schema::create('paids', function (Blueprint $table) {
       $table->id();
-      $table->integer('user_id');
-      $table->integer('payslip_id');
+      $table->bigInteger('user_id');
+      $table->bigInteger('payslip_id');
       $table->date('pay_date');
       $table->date('next_pay_date');
-      $table->integer('created_user');
-      $table->integer('updated_user');
-      $table->integer('deleted_user')->nullable();
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
       $table->softDeletes();
       $table->timestamps();
     });
