+++ database/migrations/2021_01_15_153701_create_promotions_table.php
@@ -6,35 +6,35 @@ use Illuminate\Support\Facades\Schema;
 
 class CreatePromotionsTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('promotions', function (Blueprint $table) {
-            $table->id();
-            $table->integer('schoolyear_id');
-            $table->integer('promote_schoolyear');
-            $table->integer('current_class');
-            $table->integer('promote_class');
-            $table->integer('student_id');
-            $table->integer('created_user');
-            $table->integer('updated_user');
-            $table->integer('deleted_user')->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
-    }
+  /**
+   * Run the migrations.
+   *
+   * @return void
+   */
+  public function up()
+  {
+    Schema::create('promotions', function (Blueprint $table) {
+      $table->id();
+      $table->bigInteger('schoolyear_id');
+      $table->integer('promote_schoolyear');
+      $table->integer('current_class');
+      $table->integer('promote_class');
+      $table->integer('student_id');
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
+      $table->softDeletes();
+      $table->timestamps();
+    });
+  }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('promotions');
-    }
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  public function down()
+  {
+    Schema::dropIfExists('promotions');
+  }
 }
