+++ database/migrations/2021_01_15_154748_create_studentexams_table.php
@@ -6,34 +6,34 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateStudentexamsTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('studentexams', function (Blueprint $table) {
-            $table->integer('exam_id');
-            $table->integer('student_id');
-            $table->double('total');
-            $table->double('avg');
-            $table->boolean('ispass');
-            $table->integer('created_user');
-            $table->integer('updated_user');
-            $table->integer('deleted_user')->nullable();
-            $table->softDeletes();
-            $table->timestamps();
-        });
-    }
+  /**
+   * Run the migrations.
+   *
+   * @return void
+   */
+  public function up()
+  {
+    Schema::create('studentexams', function (Blueprint $table) {
+      $table->bigInteger('exam_id');
+      $table->bigInteger('student_id');
+      $table->double('total');
+      $table->double('avg');
+      $table->boolean('ispass');
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
+      $table->softDeletes();
+      $table->timestamps();
+    });
+  }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('studentexams');
-    }
+  /**
+   * Reverse the migrations.
+   *
+   * @return void
+   */
+  public function down()
+  {
+    Schema::dropIfExists('studentexams');
+  }
 }
