+++ database/migrations/2021_02_01_121337_create_pay_schoolfees_table.php
@@ -15,14 +15,14 @@ class CreatePaySchoolfeesTable extends Migration
   {
     Schema::create('pay_schoolfees', function (Blueprint $table) {
       $table->id();
-      $table->integer('student_id');
-      $table->integer('schoolfees_id');
+      $table->bigInteger('student_id');
+      $table->bigInteger('schoolfees_id');
       $table->integer('paid_amount');
       $table->integer('left_amount');
       $table->date('paid_date');
-      $table->integer('created_user');
-      $table->integer('updated_user');
-      $table->integer('deleted_user')->nullable();
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
       $table->softDeletes();
       $table->timestamps();
     });
