+++ database/migrations/2021_02_01_121547_create_notifications_table.php
@@ -17,11 +17,11 @@ class CreateNotificationsTable extends Migration
       $table->id();
       $table->integer('title')->nullable();
       $table->string('message', 256)->nullable();
-      $table->integer('kins_id');
+      $table->bigInteger('kins_id');
       $table->boolean('is_payreminder');
-      $table->integer('created_user');
-      $table->integer('updated_user');
-      $table->integer('deleted_user')->nullable();
+      $table->bigInteger('created_user');
+      $table->bigInteger('updated_user');
+      $table->bigInteger('deleted_user')->nullable();
       $table->softDeletes();
       $table->timestamps();
     });
